+++ package.json
@@ -11,7 +11,6 @@
     "electron-log": "^4.1.1",
     "electron-store": "^5.1.1",
     "jquery": "^3.4.1",
-    "popper.js": "^1.16.1",
     "shelljs": "^0.8.3",
     "update-electron-app": "^1.5.0"
   },
