+++ src/main.ts
@@ -1,17 +1,17 @@
-import {app, BrowserWindow} from "electron";
+import { app, BrowserWindow } from "electron";
 import * as path from "path";
 
-import {log} from "./utils/LogUtil";
+import { log } from "./utils/LogUtil";
 import * as store from "./services/store/StoreManager"
 
 log.debug(`config store ->${store}`)
-log.debug(store)
+// log.debug(store)
 
-require("update-electron-app")({
-  repo: "bnh6/cryptobox",
-  updateInterval: "1 hour",
-  logger: log
-});
+// require("update-electron-app")({
+//   repo: "bnh6/cryptobox",
+//   updateInterval: "1 hour",
+//   logger: log
+// });
 
 let mainWindow: BrowserWindow;
 
@@ -36,39 +36,62 @@ function createWindow() {
   mainWindow.loadFile(path.join(__dirname, "../static/ui/index.html"));
 
   const fs = require("fs");
-  mainWindow.on("closed", function() {
+  mainWindow.on("closed", function () {
     mainWindow = null;
   });
 
-  mainWindow.on("unresponsive", function() {
+  mainWindow.on("unresponsive", function () {
     log.info("unresponsive, response to be implemented ...");
   });
 }
 
 app.on("ready", () => {
-  //loadScripts();
+  loadScripts();
   createWindow();
+
+  mainWindow.webContents.openDevTools();
 });
 
-app.on("window-all-closed", function() {
+app.on("window-all-closed", function () {
   app.quit();
 });
 
-app.on("activate", function() {
+app.on("activate", function () {
   if (mainWindow === null) createWindow();
 });
 
-process.on("uncaughtException", function(error) {
+process.on("uncaughtException", function (error) {
   console.log("UNCATCH EXCEPTION FOUND ");
   console.log(error);
 });
 
-// no longer required, also the scripts no longers sits on a single folder.
-// function loadScripts() {
-//   const scripts = fs.readdirSync("./dist/client");
-//   scripts.forEach(function(script: string) {
-//     script.endsWith(".js") &&
-//       import("./client/" + script) &&
-//       log.debug(`imported: ${script} `);
-//   });
+function loadScripts() {
+  import("./controllers/IPCManager")
+
+  //   const scripts = fs.readdirSync("./dist/client");
+  //   scripts.forEach(function(script: string) {
+  //     script.endsWith(".js") &&
+  //       import("./client/" + script) &&
+  //       log.debug(`imported: ${script} `);
+  //   });
+}
+
+
+
+// import { PasswordApplication } from "./applications/PasswordApp"
+// import { Volume } from "./entities/Volume"
+// function testing() {
+
+//   const volumeSrc: Volume = new Volume(
+//     "testingVolume",
+//     "/tmp/",
+//     null,
+//     0
+//   );
+
+//   let passwordApp = new PasswordApplication();
+//   passwordApp.checkSource(volumeSrc)
+
+//   // app.quit()
+
 // }
