+++ static/ui/index.html
@@ -10,44 +10,30 @@
 </head>
 
 <body>
-<nav class="navbar navbar-expand-md navbar-dark bg-primary">
-    <h2>CRYPTOBOX</h2>
-</nav>
-<form id="cloudEncForm">
-    <div class="form-row">
-        <div class="form-group col-md-4">
-            <label for="source">Secured Folder</label>
-            <input class="form-control" id="source" placeholder="" required type="text">
-            <div class="invalid-feedback">
-                Please provide an folder to be encrypted.
+    <nav class="navbar navbar-expand-md navbar-dark bg-primary">
+        <h2>CRYPTOBOX</h2>
+    </nav>
+    <form id="cloudEncForm">
+        <div class="form-row">
+            <div class="form-group col-md-4">
+                <label for="source">Secured Folder</label>
+                <input class="form-control" id="source" placeholder="" required type="text">
+                <div class="invalid-feedback">
+                    Please provide an folder to be encrypted.
+                </div>
             </div>
         </div>
-        <div class="form-group col-md-4">
-            <label for="destination">Working Folder</label>
-            <input class="form-control" id="destination" placeholder="" required type="text">
+
+        <div class="form-group">
+            <div class="form-group col-md-2">
+                <span id="statusLbl"></span> <br>
+                <button class="btn btn-primary" id="mountBtn" type="submit">Mount</button>
+            </div>
         </div>
-    </div>
-    <!--<div class="form-row">
-      <div class="form-group col-md-3">
-        <label for="volumeName">Volume Name</label>-->
-    <input class="form-control" id="volumeName" placeholder="VolumeName" required type="hidden" value="Cryptobox">
-    <!-- </div> -->
-    <div class="form-group col-md-3">
-        <fieldset disabled id="passwordDiv">
-            <label for="password" id="passwordLabel">Password</label>
-            <input class="form-control" id="password" placeholder="" type="password">
-        </fieldset>
-    </div>
-    </div>
-    <div class="form-group">
-        <div class="form-group col-md-2">
-            <span id="statusLbl"></span> <br>
-            <button class="btn btn-primary" id="mountBtn" type="submit">Mount</button>
-        </div>
-    </div>
-</form>
-<script>
-    require('../../src/renderer/renderer.js')
-</script>
+    </form>
+    <script>
+        require('../../src/client/volumes-renderer.js')
+    </script>
 </body>
+
 </html>
