+++ static/ui/password.html
@@ -1,0 +1,67 @@
+<body>
+    <nav class="navbar navbar-expand-md navbar-dark bg-primary">
+        <h2>CRYPTOBOX</h2>
+    </nav>
+
+    <form id="cloudEncForm">
+        <div class="form-row">
+            <div class="form-group col-md-4">
+                <label for="source">Secured Folder</label>
+                <input class="form-control" id="source" placeholder="" required type="text">
+                <div class="invalid-feedback">
+                    Please provide an folder to be encrypted.
+                </div>
+            </div>
+        
+        </div>
+        <!--<div class="form-row">
+          <div class="form-group col-md-3">
+            <label for="volumeName">Volume Name</label>-->
+        <input class="form-control" id="volumeName" placeholder="VolumeName" required type="hidden" value="Cryptobox">
+        <!-- </div> -->
+        <div class="form-group col-md-3">
+            <fieldset disabled id="passwordDiv">
+                <label for="password" id="passwordLabel">Password</label>
+                <input class="form-control" id="password" placeholder="" type="password">
+            </fieldset>
+        </div>
+        </div>
+        <div class="form-group">
+            <div class="form-group col-md-2">
+                <span id="statusLbl"></span> <br>
+                <button class="btn btn-primary" id="mountBtn" type="submit">Mount</button>
+            </div>
+        </div>
+    </form>
+    </body>
+
+
+
+    <p id="label"></p>
+    <p>
+      <input type = "text" id="password" value="">
+      <input type = "button" id="ok" value="OK" onclick="response()">
+      <input type = "button" value="Cancel" onclick="cancel()">
+    </p>
+    
+    <script>
+    const { ipcRenderer } = require("electron")
+    function cancel() {
+      ipcRenderer.send("closeDialog", "")
+      this.close();
+    }
+    
+    function response() {
+      ipcRenderer.send("closeDialog", document.getElementById("data").value)
+      this.close();
+    }
+    
+    window.onload=function() {
+      var options = ipcRenderer.sendSync("openDialog", "")
+      var params = JSON.parse(options)
+      document.getElementById("label").innerHTML = params.label;
+      document.getElementById("data").value = params.value;
+      document.getElementById("ok").value = params.ok;
+    }
+    </script>
+    </body>
